+++ internal/provider/types/request_body.go
@@ -5,6 +5,6 @@ package types
 import "github.com/hashicorp/terraform-plugin-framework/types"
 
 type RequestBody struct {
-	MaxSize     types.Int64 `tfsdk:"max_size"`
+	MaxSize     types.Int32 `tfsdk:"max_size"`
 	SendRawBody types.Bool  `tfsdk:"send_raw_body"`
 }
